New Features
added additional features to the function "makeActive". It will now also assign class "active-button" to the navigation bar buttons as the user scrolls through the sections to show which section is the user on currently.

Added "hideNavBar" function to make sure the navigation bar is hidden if the user is not scrolling.

Added the "scrollAction" function to provide smoother scrolling to the specific section when clicked.

Added a "Return to Top" button at the bottom of the page right above the footer so the user doesn't have to scroll all the way up manually.

Added active, active_button, return_to_top classes to CSS file